컨벤션 재수정

+++ src/main/java/shop/kimkj/mytrip/service/UserReviewService.java
@@ -4,7 +4,10 @@ package shop.kimkj.mytrip.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import shop.kimkj.mytrip.domain.UserReview;
+import shop.kimkj.mytrip.dto.UserReviewCommentRequestDto;
 import shop.kimkj.mytrip.dto.UserReviewRequestDto;
 import shop.kimkj.mytrip.repository.UserReviewRepository;
 
@@ -24,8 +27,8 @@ public class UserReviewService {
         return userReview;
     }
 
-    public UserReview getUserReview(Long reviewId) {
-        return userReviewRepository.findById(reviewId).orElseThrow(
+    public UserReview getUserReview(int reviewId) {
+        return userReviewRepository.findById((long)reviewId).orElseThrow(
                 () -> new NullPointerException("해당 리뷰가 존재하지 않습니다.")
         );
     }
@@ -35,12 +38,12 @@ public class UserReviewService {
     }
 
     @Transactional
-    public String deleteUserReview(Long reviewId) {
-        userReviewRepository.deleteById(reviewId);
+    public String deleteUserReview(int reviewId) {
+        userReviewRepository.deleteById((long)reviewId);
         return "삭제를 완료했습니다.";
     }
 
-//    public UserReview postUserReviewComment(@PathVariable Long id,@RequestBody UserReviewCommentRequestDto userReviewCommentRequestDto) {
+//    public UserReview postUserReviewComment(@PathVariable Long id, @RequestBody UserReviewCommentRequestDto userReviewCommentRequestDto) {
 //        Comment userReview = new Comment(userReviewCommentRequestDto);
 //        userReviewRepository.save(userReview);
 //        return userReview;
